brand

+++ backend/routers/brands.py
@@ -1,7 +1,8 @@
-import os
+# coding:utf-8
 from datetime import datetime
 from fastapi import APIRouter, Depends, UploadFile
 from db.db import get_db
+from schemas.product import Products
 from models.brand import BrandCreate, BrandUpdate, BrandResponse
 from schemas.brand import Brands
 from models.common import BaseResponse
@@ -34,7 +35,7 @@ async def add_brand(logo: Optional[UploadFile], brand: BrandCreate = Depends(Bra
         # 获取当前路由位置作为文件路径
         full_name = generate_filename(
             prefix=router.prefix, filename=logo.filename, name=None)
-        print(full_name)
+        # print(full_name)
         oss_client = AliyunOSS(
             endpoint=endpoint, region=region, bucket_name=bucket_name)
         res = await oss_client.upload_file(
@@ -57,20 +58,25 @@ async def add_brand(logo: Optional[UploadFile], brand: BrandCreate = Depends(Bra
 
 @router.delete('/delete/{id}')
 async def delete_brand(id: int, db: Session = Depends(get_db)):
-    db_item = db.query(Brands).filter_by(id=id).first()
+    # check_res = await check_brand_id(id)
+    check_res = db.query(Products).filter_by(brand_id=id).first()
+    if not check_res:
+        db_item = db.query(Brands).filter_by(id=id).first()
 
-    if not db_item:
-        return BaseResponse.error(code=1, message="品牌不存在")
+        if not db_item:
+            return BaseResponse.error(code=1, message="品牌不存在")
 
-    res = db.delete(db_item)
-    db.commit()
-    return BaseResponse.success(data=res)
+        res = db.delete(db_item)
+        db.commit()
+        return BaseResponse.success(data=res)
+    else:
+        return BaseResponse.error(code=1, message="已存在产品关联该品牌")
 
 # table 需要加一个image_id key
 
 
 @router.post('/update')
-async def update_brand(logo: Optional[UploadFile], brand: BrandUpdate, db: Session = Depends(get_db)):
+async def update_brand(logo: Optional[UploadFile], brand: BrandUpdate = Depends(BrandUpdate.as_form), db: Session = Depends(get_db)):
     db_item = db.query(Brands).filter_by(id=brand.id).first()
     if not db_item:
         return BaseResponse.error(code=1, message="品牌不存在")
@@ -80,16 +86,18 @@ async def update_brand(logo: Optional[UploadFile], brand: BrandUpdate, db: Sessi
         if logo.content_type not in allowed_types:
             return BaseResponse.error(code=1, message="上传文件内容不允许")
         raw = await logo.read()
-        oss_client = AliyunOSS(
-            endpoint=endpoint, region=region, bucket_name=bucket_name)
-        full_name = generate_filename(prefix=router.prefix,
-                                      filename=logo.filename, name='logo')
-        res = await oss_client.upload_file(
-            name=full_name, data=raw)
-        db_item.logo_url = res.get('url') if res.get('url') else ''
-        db_item.key = res.get('key') if res.get('key') else ''
-        db_item.file_id = res.get('file_id') if res.get('file_id') else ''
-
+        if raw:
+            oss_client = AliyunOSS(
+                endpoint=endpoint, region=region, bucket_name=bucket_name)
+            full_name = generate_filename(
+                prefix=router.prefix, filename=logo.filename, name=None)
+            res = await oss_client.upload_file(
+                name=full_name, data=raw)
+            db_item.logo_url = res.get('url') if res.get('url') else ''
+            db_item.key = res.get('key') if res.get('key') else ''
+            db_item.file_id = res.get('file_id') if res.get('file_id') else ''
+        else:
+            return BaseResponse.error(code=1, message='图片上传失败')
     db_item.description = brand.description
     db_item.is_deleted = brand.is_deleted
     db_item.name = brand.name
